separate Arithmetic
