+++ src/main/java/org/modelingvalue/dclare/Logic.java
@@ -23,7 +23,6 @@ package org.modelingvalue.dclare;
 import java.lang.reflect.InvocationHandler;
 import java.lang.reflect.Method;
 import java.lang.reflect.Proxy;
-import java.math.BigInteger;
 import java.util.Objects;
 import java.util.function.BiFunction;
 
@@ -43,10 +42,9 @@ import org.modelingvalue.collections.util.SerializableSupplier;
 import org.modelingvalue.collections.util.SerializableSupplier.SerializableSupplierImpl;
 import org.modelingvalue.collections.util.SerializableTriFunction;
 import org.modelingvalue.collections.util.SerializableTriFunction.SerializableTriFunctionImpl;
+import org.modelingvalue.dclare.Arithmetic.Pred;
 
-public final class Logic {
-    private Logic() {
-    }
+public class Logic {
 
     private static final boolean                                              USE_EXTEND = Boolean.getBoolean("USE_EXTEND");
 
@@ -161,7 +159,7 @@ public final class Logic {
     }
 
     @SuppressWarnings("rawtypes")
-    private static final Object unproxy(Object object) {
+    protected static final Object unproxy(Object object) {
         if (object instanceof Term) {
             return Proxy.getInvocationHandler(object);
         } else {
@@ -949,68 +947,10 @@ public final class Logic {
         return term(INCOMPLETE_FUNCTOR_PROXY, der);
     }
 
-    // Facts, Is
-
-    public static void fact(Term term) {
-        ((TermImpl<?>) unproxy(term)).makeFact();
-    }
-
-    // Variable bindings
+    // Equals
 
     @SuppressWarnings({"unchecked", "rawtypes"})
-    public static Map<Variable, Object> bind(Term... varVal) {
-        Map<Variable, Object> b = Map.of();
-        for (int i = 0; i < varVal.length; i += 2) {
-            b = b.add(Entry.of((Variable) varVal[i], varVal[i + 1]));
-        }
-        return b;
-    }
-
-    // Is
-
-    private static Functor<Pred> is = functor((SerializableBiFunction<Int, IntLit, Pred>) Logic::is);
-
-    public static Pred is(Int i, IntLit r) {
-        return term(is, i, r);
-    }
-
-    // Integer
-
-    public static interface Int extends Term {
-    }
-
-    public static interface IntLit extends Int {
-    }
-
-    public static interface IntFun extends Int {
-    }
-
-    private static Functor<IntLit> i = functor((SerializableFunction<BigInteger, IntLit>) Logic::i);
-
-    private static IntLit i(BigInteger x) {
-        return term(i, x);
-    }
-
-    public static IntLit i(long x) {
-        return i(BigInteger.valueOf(x));
-    }
-
-    public static IntLit ilv(String name) {
-        return var(IntLit.class, name);
-    }
-
-    public static IntFun ifv(String name) {
-        return var(IntFun.class, name);
-    }
-
-    public static Int iv(String name) {
-        return var(Int.class, name);
-    }
-
-    // Eq
-
-    @SuppressWarnings({"unchecked", "rawtypes"})
-    private static Functor<Pred> eq = functor(Logic::eq, t -> {
+    private static Functor<Pred> eq = functor(Arithmetic::eq, t -> {
         TermImpl at = t.getTerm(1);
         TermImpl bt = t.getTerm(2);
         if (at == null && bt == null) {
@@ -1028,138 +968,21 @@ public final class Logic {
         return term(eq, a, b);
     }
 
-    // Operators
+    // Facts, Is
 
-    public static interface Pred extends Term {
+    public static void fact(Term term) {
+        ((TermImpl<?>) unproxy(term)).makeFact();
     }
+
+    // Variable bindings
 
     @SuppressWarnings({"unchecked", "rawtypes"})
-    private static Functor<Pred> plusPred = functor((SerializableTriFunction<IntLit, IntLit, IntLit, Pred>) Logic::plus, t -> {
-        TermImpl<IntLit> at = t.getTerm(1);
-        TermImpl<IntLit> bt = t.getTerm(2);
-        TermImpl<IntLit> ct = t.getTerm(3);
-        BigInteger ai = at != null ? at.getVal(1) : null;
-        BigInteger bi = bt != null ? bt.getVal(1) : null;
-        BigInteger ci = ct != null ? ct.getVal(1) : null;
-        if (ai != null && bi != null && ci != null) {
-            return ai.add(bi).equals(ci) ? Set.of(t) : Set.of();
-        } else if (ai != null && bi != null && ci == null) {
-            return Set.of(t.set(3, at.set(1, ai.add(bi))));
-        } else if (ai != null && bi == null && ci != null) {
-            return Set.of(t.set(2, at.set(1, ci.subtract(ai))));
-        } else if (ai == null && bi != null && ci != null) {
-            return Set.of(t.set(1, bt.set(1, ci.subtract(bi))));
-        } else {
-            return t.incomplete();
+    public static Map<Variable, Object> bind(Term... varVal) {
+        Map<Variable, Object> b = Map.of();
+        for (int i = 0; i < varVal.length; i += 2) {
+            b = b.add(Entry.of((Variable) varVal[i], varVal[i + 1]));
         }
-    });
-
-    public static Pred plus(IntLit a, IntLit b, IntLit r) {
-        return term(plusPred, a, b, r);
-    }
-
-    @SuppressWarnings({"unchecked", "rawtypes"})
-    private static Functor<Pred> multiplyPred = functor((SerializableTriFunction<IntLit, IntLit, IntLit, Pred>) Logic::multiply, t -> {
-        TermImpl<IntLit> at = t.getTerm(1);
-        TermImpl<IntLit> bt = t.getTerm(2);
-        TermImpl<IntLit> ct = t.getTerm(3);
-        BigInteger ai = at != null ? at.getVal(1) : null;
-        BigInteger bi = bt != null ? bt.getVal(1) : null;
-        BigInteger ci = ct != null ? ct.getVal(1) : null;
-        if (ai != null && bi != null && ci != null) {
-            return ai.multiply(bi).equals(ci) ? Set.of(t) : Set.of();
-        } else if (ai != null && bi != null && ci == null) {
-            return Set.of(t.set(3, at.set(1, ai.multiply(bi))));
-        } else if (ai != null && bi == null && ci != null) {
-            return Set.of(t.set(2, at.set(1, ci.divide(ai))));
-        } else if (ai == null && bi != null && ci != null) {
-            return Set.of(t.set(1, bt.set(1, ci.divide(bi))));
-        } else {
-            return t.incomplete();
-        }
-    });
-
-    public static Pred multiply(IntLit a, IntLit b, IntLit r) {
-        return term(multiplyPred, a, b, r);
-    }
-
-    @SuppressWarnings({"unchecked", "rawtypes"})
-    private static Functor<Pred> powerPred = functor((SerializableBiFunction<IntLit, IntLit, Pred>) Logic::power, t -> {
-        TermImpl<IntLit> at = t.getTerm(1);
-        TermImpl<IntLit> bt = t.getTerm(2);
-        BigInteger ai = at != null ? at.getVal(1) : null;
-        BigInteger bi = bt != null ? bt.getVal(1) : null;
-        if (ai != null && bi != null) {
-            return ai.multiply(ai).equals(bi) ? Set.of(t) : Set.of();
-        } else if (ai != null && bi == null) {
-            return Set.of(t.set(2, at.set(1, ai.multiply(ai))));
-        } else if (ai == null && bi != null) {
-            BigInteger sqrt = bi.sqrt();
-            return Set.of(t.set(1, bt.set(1, sqrt)), t.set(1, bt.set(1, sqrt.negate())));
-        } else {
-            return t.incomplete();
-        }
-    });
-
-    public static Pred power(IntLit a, IntLit r) {
-        return term(powerPred, a, r);
-    }
-
-    // Functions
-
-    private static Functor<IntFun> plusFunc = functor((SerializableBiFunction<Int, Int, IntFun>) Logic::plus);
-
-    public static IntFun plus(Int a, Int b) {
-        return term(plusFunc, a, b);
-    }
-
-    private static Functor<IntFun> minusFunc = functor((SerializableBiFunction<Int, Int, IntFun>) Logic::minus);
-
-    public static IntFun minus(Int a, Int b) {
-        return term(minusFunc, a, b);
-    }
-
-    private static Functor<IntFun> multiplyFunc = functor((SerializableBiFunction<Int, Int, IntFun>) Logic::multiply);
-
-    public static IntFun multiply(Int a, Int b) {
-        return term(multiplyFunc, a, b);
-    }
-
-    private static Functor<IntFun> divideFunc = functor((SerializableBiFunction<Int, Int, IntFun>) Logic::divide);
-
-    public static IntFun divide(Int a, Int b) {
-        return term(divideFunc, a, b);
-    }
-
-    private static Functor<IntFun> powerFunc = functor((SerializableFunction<Int, IntFun>) Logic::power);
-
-    public static IntFun power(Int a) {
-        return term(powerFunc, a);
-    }
-
-    private static Functor<IntFun> sqrtFunc = functor((SerializableFunction<Int, IntFun>) Logic::sqrt);
-
-    public static IntFun sqrt(Int a) {
-        return term(sqrtFunc, a);
-    }
-
-    // Is Rules
-
-    public static void isRules() {
-        IntLit PL = ilv("PL");
-        IntLit QL = ilv("QL");
-        IntLit RL = ilv("RL");
-
-        Int X = iv("X");
-        Int Y = iv("Y");
-
-        rule(is(PL, RL), eq(PL, RL));
-        rule(is(plus(X, Y), RL), is(X, PL), is(Y, QL), plus(PL, QL, RL));
-        rule(is(minus(X, Y), RL), is(X, PL), is(Y, QL), plus(RL, QL, PL));
-        rule(is(multiply(X, Y), RL), is(X, PL), is(Y, QL), multiply(PL, QL, RL));
-        rule(is(divide(X, Y), RL), is(X, PL), is(Y, QL), multiply(RL, QL, PL));
-        rule(is(power(X), RL), is(X, PL), power(PL, RL));
-        rule(is(sqrt(X), RL), is(X, PL), power(RL, PL));
+        return b;
     }
 
 }
